# module-dto/src/main/java/org/example/moduledto/exception/validation/ValidationErrorResponse.java
package org.example.moduledto.exception.validation;

import lombok.Getter;
import lombok.RequiredArgsConstructor;

import java.util.List;

@Getter
@RequiredArgsConstructor
public class ValidationErrorResponse {

    private final List<Violation> violations;

}