# module-dto/src/main/java/org/example/moduledto/exception/validation/Violation.java
package org.example.moduledto.exception.validation;

import lombok.Getter;
import lombok.RequiredArgsConstructor;

@Getter
@RequiredArgsConstructor
public class Violation {

    private final String fieldName;
    private final String message;

}